Move footer to components folder
