+++ personal-website/src/app/blog/page.tsx
@@ -1,6 +1,6 @@
 import Blog from "../../components/blog";
 import Navbar from "../../components/navbar";
-import Footer from "../footer";
+import Footer from "../../components/footer";
 
 export default function BlogPage() {
   type Blog = {
