+++ personal-website/src/app/contact/page.tsx
@@ -1,5 +1,5 @@
 import Navbar from "../../components/navbar";
-import Footer from "../footer";
+import Footer from "../../components/footer";
 
 export default function ContactPage() {
   return (
