+++ personal-website/src/app/page.tsx
@@ -1,6 +1,6 @@
 import Image from "next/image";
 import Navbar from "../components/navbar";
-import Footer from "./footer";
+import Footer from "../components/footer";
 
 export default function HomePage() {
   return (
