+++ personal-website/src/app/resume/page.tsx
@@ -1,6 +1,6 @@
 import Link from "next/link";
 import Navbar from "../../components/navbar";
-import Footer from "../footer";
+import Footer from "../../components/footer";
 
 export default function ResumePage() {
   return (
